@@ -54,9 +54,9 @@
   "devDependencies": {
     "babel": "^5.1.13",
     "babel-core": "^5.1.13",
-    "babel-eslint": "^3.1.7",
-    "eslint": "^0.21.2",
-    "eslint-plugin-react": "^2.5.2",
+    "babel-eslint": "^4.0.5",
+    "eslint": "^1.0.0",
+    "eslint-plugin-react": "^3.2.0",
     "tomchentw-npm-dev": "^3.0.0"
   },
   "dependencies": {
