@@ -1,9 +1,9 @@
-import React from "react";
+import {default as React, Component} from "react";
 import {default as GitHubForkRibbon} from "react-github-fork-ribbon";
 import {ToastContainer, ToastMessage} from "react-toastr";
 
-import NavHeaderBar from "./NavHeaderBar";
-import ComponentPlayground from "./ComponentPlayground";
+import {default as NavHeaderBar} from "./NavHeaderBar";
+import {default as ComponentPlayground} from "./ComponentPlayground";
 
 const ACTIONS = [
   {
@@ -50,17 +50,6 @@ const DROPDOWN_ACTIONS = [
       componentClass: require("./components/basics/Geolocation"),
       componentRaw: {
         __raw: require("!raw-loader!./components/basics/Geolocation"),
-      },
-    },
-  },
-  {
-    key: "basics__asynchronous-loading",
-    displayName: "Asynchronous Loading",
-    path: "#basics/asynchronous-loading",
-    component: {
-      componentClass: require("./components/basics/AsynchronousLoading"),
-      componentRaw: {
-        __raw: require("!raw-loader!./components/basics/AsynchronousLoading"),
       },
     },
   },
@@ -164,7 +153,7 @@ const RIGHT_ACTIONS = [
 
 const ALL_ACTIONS = ACTIONS.concat(DROPDOWN_ACTIONS.filter((x) => { return !!x; })).concat(RIGHT_ACTIONS);
 
-class ReactRoot extends React.Component {
+export default class ReactRoot extends Component {
 
   constructor (...args) {
     super(args);
@@ -228,5 +217,3 @@ class ReactRoot extends React.Component {
     );
   }
 }
-
-export default ReactRoot;
