@@ -1,35 +1,31 @@
-import React from "react/addons";
-import {GoogleMaps} from "react-google-maps";
+import {default as React, Component} from "react";
 
+import {default as GoogleMap} from "../../../../../src/GoogleMap";
 /*
  * Sample From: https://developers.google.com/maps/documentation/javascript/examples/map-simple
+ *
+ * Add <script src="https://maps.googleapis.com/maps/api/js"></script> to your HTML to provide google.maps reference
  */
-class SimpleMap extends React.Component {
+export default class SimpleMap extends Component {
   /*
    * 1. Create a component that wraps all your map sub-components.
    */
   render () {
-    const {props, state} = this,
-          {googleMapsApi, ...otherProps} = props;
     /*
-     * 2. Render GoogleMaps component with containerProps
+     * 2. Render GoogleMap component with containerProps
      */
     return (
-      <GoogleMaps containerProps={{
-          ...otherProps,
+      <GoogleMap containerProps={{
+          ...this.props,
           style: {
             height: "100%",
           },
         }}
         /*
-         * 3. Pass googleMapsApi from global variable as props.
+         * 3. config <GoogleMap> instance by properties
          */
-        googleMapsApi={google.maps}
-        zoom={8}
-        center={{lat: -34.397, lng: 150.644}} />
+        defaultZoom={8}
+        defaultCenter={{lat: -34.397, lng: 150.644}} />
     );
   }
-
 }
-
-export default SimpleMap;
