@@ -1,20 +1,21 @@
-import React from "react/addons";
-import {GoogleMaps, InfoWindow, Marker} from "react-google-maps";
+import {default as React, Component} from "react";
+
+import {default as GoogleMap} from "../../../../../src/GoogleMap";
+import {default as InfoWindow} from "../../../../../src/InfoWindow";
 
 /*
  * https://developers.google.com/maps/documentation/javascript/examples/event-properties
+ *
+ * Add <script src="https://maps.googleapis.com/maps/api/js"></script> to your HTML to provide google.maps reference
  */
-class GettingProperties extends React.Component {
+export default class GettingProperties extends Component {
 
-  constructor (...args) {
-    super(...args);
-    this.state = {
-      zoomLevel: 4,
-      content: "Change the zoom level", 
-    };
+  state = {
+    zoomLevel: 4,
+    content: "Change the zoom level", 
   }
 
-  _handle_zoom_changed () {
+  _handle_zoom_changed = () => {
     const zoomLevel = this.refs.map.getZoom();
     if (zoomLevel !== this.state.zoomLevel) {
       // Notice: Check zoomLevel equality here,
@@ -27,27 +28,25 @@ class GettingProperties extends React.Component {
   }
 
   render () {
-    const {props, state} = this,
-          {googleMapsApi, ...otherProps} = props,
-          myLatLng = new google.maps.LatLng(-25.363882, 131.044922);
+    const myLatLng = new google.maps.LatLng(-25.363882, 131.044922);
+    const {zoomLevel, content} = this.state;
 
     return (
-      <GoogleMaps containerProps={{
-          ...otherProps,
+      <GoogleMap containerProps={{
+          ...this.props,
           style: {
             height: "100%",
           },
         }}
         ref="map"
-        googleMapsApi={google.maps}
-        zoom={state.zoomLevel}
-        center={myLatLng}
-        onZoomChanged={this._handle_zoom_changed.bind(this)}>
-        <InfoWindow position={myLatLng} content={state.content} />
-      </GoogleMaps>
+        defaultCenter={myLatLng}
+        zoom={zoomLevel}
+        onZoomChanged={this._handle_zoom_changed}>
+        <InfoWindow
+          defaultPosition={myLatLng}
+          content={content}
+        />
+      </GoogleMap>
     );
   }
-
 }
-
-export default GettingProperties;
